server/requirements.txt
asgiref==3.11.0
Django==6.0
django-stubs
djangorestframework-stubs
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
redis>=4.5.2,<5.0
django-redis==5.4.0
celery[redis]==5.3.4
python-dotenv==1.0.0
sqlparse==0.5.5
django-extensions==4.1
pytest==9.0.2
pytest-django==4.11.1